+++ Runtime/Scheduler/AssemblyInfo.cs
@@ -1,6 +1,7 @@
 using System.Runtime.CompilerServices;
 
 [assembly: InternalsVisibleTo("Unity.Services.Core.Registration")]
+[assembly: InternalsVisibleTo("Unity.Services.Core.TestUtils")]
 
 #if UNITY_INCLUDE_TESTS
 [assembly: InternalsVisibleTo("Unity.Services.Core.Tests")]