+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.1.0-pre.11] - 2021-10-22
+### Added
+- Getter methods for `ConfigurationBuilder`.
+
+### Fixed
+- Fix layout for Project Bind Redirect Popup for Light theme
+
 ## [1.1.0-pre.10] - 2021-10-08
 ### Added
 - `IActionScheduler` component to schedule actions at runtime.